src-python/pyproject.toml
[project]
name = "doc-anonymizer-sidecar"
version = "0.1.0"
description = "Offline document anonymizer - Python sidecar backend"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pypdfium2>=4.30.0",
    "Pillow>=11.0.0",
    "pytesseract>=0.3.13",
    "spacy>=3.8.0",
    "llama-cpp-python>=0.3.0",
    "cryptography>=44.0.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]
office = [
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
